LayoutTests/html5lib/run-test7-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '<'
CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '<'
resources/tests7.dat: PASS

Switching to data: url parser (threaded if available) and re-running...

resources/tests7.dat: PASS
